feat: scoped JWT auth + refactor app.py
JWT now carries a 'scope' claim ('bench' for full access, 'site' for site-scoped). has_scope() enforces site restrictions. require_scope() decorator applied to all 22 site routes in sites.py. Bearer token support in Authorization header.

app.py refactored from ~720 lines to ~250:
- Bench routes extracted to views/benches.py
- Rate limiting extracted to rate_limit.py
- Bench helpers extracted to bench_helpers.py
- Auth decorators extracted to auth.py

Route: /api/benches/<name>/<action> -> /api/benches/<name>/actions/<action_name>
